LoggerFactory moved to Utility namespace
